feat(variables): move the variables component in website
