+++ packages/mantine/src/index.ts
@@ -12,7 +12,6 @@ export * from '@mantine/notifications';
 export {type NotificationsProps} from '@mantine/notifications';
 export * from '@tanstack/table-core';
 export * from './components';
-export {CssVariablesList} from './utils';
 export {noop};
 
 // explicitly overriding mantine components
