+++ packages/mantine/src/theme/index.ts
@@ -1,1 +1,2 @@
+export * from './plasmaCSSVariablesResolver';
 export * from './Plasmantine';
