+++ packages/mantine/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './createFactoryComponent';
 export * from './createPolymorphicComponent';
-export * from './CssVariablesList';
 export * from './overrideComponent';
